requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-contrib-comments==1.9.0
django-crispy-forms==1.7.2
django-haystack==2.8.1
django-markdown-deux==1.0.5
django-mdeditor==0.1.9
django-pure-pagination==0.3.0
django-widget-tweaks==1.4.3
idna==2.7
jieba==0.39
Markdown==3.0.1
markdown2==2.3.7
mysqlclient==1.3.13
oauthlib==2.1.0
Pillow==5.3.0
pycparser==2.19
Pygments==2.2.0
PyMySQL==0.9.2
python3-openid==3.1.0
pytz==2018.6
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.1
Whoosh==2.7.4
